FlaskWebProject3/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
appdirs==1.4.4
APScheduler==3.10.4
astroid==3.3.5
asttokens>=2.1.0
attrs==23.2.0
Automat==22.10.0
azure-core>=1.30.0
azure-storage-blob==12.24.1
backcall==0.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bidict==0.23.1
blake3==1.0.4
bleach==6.1.0
blin==0.2.1
blinker==1.7.0
boto3==1.37.9
botocore==1.37.9
Brotli==1.1.0
cachelib==0.9.0
cachetools==5.3.3
captcha==0.6.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloc==0.2.5
colorama==0.4.6
colored==2.3.1
consonance==0.1.2
constantly==23.10.4
cryptography==45.0.3
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
dissononce==0.34.3
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.15.4
flask_session_captcha==1.4.2
Flask==3.0.2
Flask-APScheduler==1.13.1
Flask-Caching==2.3.0
Flask-Compress==1.14
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Session==0.6.0
Flask-Sessionstore==0.4.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flaskwebgui==1.1.3
Flask-WTF==1.2.1
fpdf==1.7.2
frozenlist==1.4.1
google-api-core==2.18.0
google-api-python-client==2.125.0
googleapis-common-protos==1.63.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
ifaddr==0.2.0
incremental==22.10.0
inflate64==1.0.1
ipython==8.12.3
isodate==0.7.2
isort==5.13.2
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.2.2
mando==0.7.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
modulegraph==0.19.6
multidict==6.0.5
multivolumefile==0.2.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
netifaces==0.11.0
Nuitka==2.1.4
oauth2client==4.1.3
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.2
pandocfilters==1.5.1
paramiko==3.5.1
parsel==1.9.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
pillow==10.4.0
pip==25.0.1
pipreqs==0.5.0
platformdirs==4.3.6
ply==3.11
prompt_toolkit==3.0.48
Protego==0.3.1
protobuf==4.25.3
proto-plus==1.23.0
psutil==5.9.8
pure_eval==0.2.3
puresnmp==2.0.0.post1
py7zr==0.22.0
pyasn1_modules==0.4.0
pyasn1==0.6.0
pybcj==1.0.3
pyclean==3.0.0
pycparser==2.22
py-cpuinfo==9.0.0
pycryptodome==3.20.0
pycryptodomex==3.22.0
PyDispatcher==2.0.7
PyDrive==1.3.1
Pygments==2.18.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.0
PyJWT==2.8.0
pylint==3.3.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.1.2
pyppmd==1.1.1
pysmb==1.2.9.1
pysmi==1.4.4
pysnmp==7.1.16
pyspnego==0.11.2
py-spy==0.3.14
pystray==0.19.5
pyte==0.8.2
python-axolotl==0.2.2
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.9.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-magic==0.4.27
python-magic-bin==0.4.14
python-socketio==5.11.3
pytsk3==20231007
pytz==2024.2
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
pyzstd==0.16.2
queuelib==1.7.0
radon==6.0.1
referencing==0.35.1
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.20.0
rsa==4.9
s3transfer==0.11.4
scapy==2.5.0
schedule==1.2.1
scheduler==0.8.5
Scrapy==2.11.2
service-identity==24.1.0
setuptools==65.5.0
shadowcopy==0.0.4
simple-websocket==1.0.0
six==1.10
smbprotocol==1.15.0
soupsieve==2.6
SQLAlchemy==2.0.27
sqlalchemy-json==0.7.0
sspilib==0.3.1
stack-data==0.6.3
t61codec==1.0.1
texttable==1.7.0
tinycss2==1.3.0
tldextract==5.1.2
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transitions==0.9.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typeguard==4.1.5
typing_extensions==4.10.0
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
w3lib==2.2.1
wait-response==1.0.0
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.42.0
winshell==0.6
WMI==1.5.1
wsproto==1.2.0
WTForms==3.1.2
x690==1.0.0.post1
XlsxWriter==3.2.0
xxhash==3.5.0
yarg==0.1.9
yarl==1.9.4
yowsup2==3.2.3
zeroconf==0.131.0
zope.interface==6.4.post2
zstandard==0.23.0
zstd==1.5.6.1  
xlsxwriter
flask-talisman
